+++ src/main/java/com/project/bookstudy/studygroup/domain/Payment.java
@@ -23,16 +23,16 @@ public class Payment {
     private LocalDateTime paymentDt;
     @Enumerated(EnumType.STRING)
     private PaymentStatus status;
-    private int price;
-    private int discountPrice;
-    private int paymentPrice;
+    private Long price;
+    private Long discountPrice;
+    private Long paymentPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder(access = AccessLevel.PRIVATE)
-    private Payment(int price, int discountPrice, int paymentPrice, Member member) {
+    private Payment(Long price, Long discountPrice, Long paymentPrice, Member member) {
         this.price = price;
         this.discountPrice = discountPrice;
         this.paymentPrice = paymentPrice;
@@ -46,7 +46,7 @@ public class Payment {
 
         //나중에 할인 정책도입시 할인 금액 계산 로직 작성
 
-        int price = studyGroup.getPrice();
+        Long price = studyGroup.getPrice();
         member.usePoint(price); // IllegalStateException
 
         Payment payment = Payment.builder()
@@ -58,9 +58,9 @@ public class Payment {
         return payment;
     }
 
-    public int refund() {
+    public Long refund() {
         if (status == PaymentStatus.REFUNDED) {
-            return 0;
+            return 0L;
         }
 
         this.status = PaymentStatus.REFUNDED;
