+++ src/main/java/com/project/bookstudy/studygroup/domain/param/CreateStudyGroupParam.java
@@ -12,14 +12,14 @@ public class CreateStudyGroupParam {
     private String contents;
     private String contentsDetail;
     private int maxSize;
-    private int price;
+    private Long price;
     private LocalDateTime studyStartDt;
     private LocalDateTime studyEndDt;
     private LocalDateTime recruitmentStartDt;
     private LocalDateTime recruitmentEndDt;
 
     @Builder
-    public CreateStudyGroupParam(String subject, String contents, String contentsDetail, int maxSize, int price, LocalDateTime studyStartDt, LocalDateTime studyEndDt, LocalDateTime recruitmentStartDt, LocalDateTime recruitmentEndDt) {
+    public CreateStudyGroupParam(String subject, String contents, String contentsDetail, int maxSize, Long price, LocalDateTime studyStartDt, LocalDateTime studyEndDt, LocalDateTime recruitmentStartDt, LocalDateTime recruitmentEndDt) {
         this.subject = subject;
         this.contents = contents;
         this.contentsDetail = contentsDetail;
